requirements.txt
numpy
requests
urllib3
python-dateutil
polygon-api-client
alpaca-py
